Move configuration management to _conf_for():
This change was made to work around the fact that we cannot call
plugin_settings() lazy later when already in the request cycle
because he will not be able to find the DSL object in the caller stack, or
find the wrong one.

It also fixed the problem that multiple apps in the same Perl process
would share the configuration for the plugin.

We call the plugin_setting early now, when we process the 'need' keyword
during setup, and we cache the config per App class name.

We could bring back the lazy features of the previous code if/when the
Dancer2::Plugin API changes to accept a $dsl parameter
on plugin_settings().

Signed-off-by: Pedro Melo <melo@simplicidade.org>